+++ features/handler/product.go
@@ -5,6 +5,7 @@ import (
 	"lokasani/entity/request"
 	"lokasani/entity/response"
 	"lokasani/features/services"
+	"lokasani/helpers/middleware"
 
 	"github.com/labstack/echo/v4"
 )
@@ -18,6 +19,14 @@ func NewProductHandler(iProductService services.IProductService) *ProductHandler
 }
 
 func (pr *ProductHandler) CreateProduct(c echo.Context) error {
+	_, roleId, _, err := middleware.ExtractToken(c)
+    if err != nil {
+        return response.NewErrorResponse(c, err)
+    }
+    if roleId != 1 {
+		return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	}
+
 	var input request.Product
 	c.Bind(&input)
 	file, err := c.FormFile("file")
@@ -100,6 +109,14 @@ func (pr *ProductHandler) GetProduct(c echo.Context) error {
 }
 
 func (pr *ProductHandler) UpdateProduct(c echo.Context) error {
+	_, roleId, _, err := middleware.ExtractToken(c)
+    if err != nil {
+        return response.NewErrorResponse(c, err)
+    }
+    if roleId != 1 {
+		return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	}
+
 	id := c.Param("id")
 	var input request.Product
 	c.Bind(&input)
@@ -112,6 +129,14 @@ func (pr *ProductHandler) UpdateProduct(c echo.Context) error {
 }
 
 func (pr *ProductHandler) DeleteProduct(c echo.Context) error {
+	_, roleId, _, err := middleware.ExtractToken(c)
+    if err != nil {
+        return response.NewErrorResponse(c, err)
+    }
+    if roleId != 1 {
+		return response.NewErrorResponse(c, echo.ErrUnauthorized)
+	}
+	
 	id := c.Param("id")
 	res, err := pr.productService.DeleteProduct(id)
 	if err != nil {
